requirements.txt
awsebcli==3.4.6
bcrypt==1.1.1
boto==2.38.0
cement==2.4.0
certifi==14.5.14
cffi==1.1.0
dj-database-url==0.3.0
Django==1.8.2
django-axes==1.3.9
django-nocaptcha-recaptcha==0.0.18
django-phonenumber-field==0.7.2
django-storages==1.1.8
django-widget-tweaks==1.3
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
Flask==0.10.1
gunicorn==19.3.0
image==1.4.0
ipython==3.0.0
itsdangerous==0.24
Jinja2==2.7.3
jmespath==0.7.1
jsonschema==2.4.0
MarkupSafe==0.23
matplotlib==1.4.3
mavproxy==1.4.23
mistune==0.5.1
numpy==1.8.1
pandas==0.16.0
pathspec==0.3.3
phonenumbers==7.0.4
Pillow==2.9.0
ptyprocess==0.4
pycparser==2.13
Pygments==2.0.2
pymavlink==1.1.57
pyparsing==2.0.3
pyreadline==2.0
pyserial==2.7
python-dateutil==2.4.1
pytz==2015.2
PyYAML==3.11
pyzmq==14.5.0
requests==2.6.2
requirements==0.1
six==1.9.0
SQLAlchemy==1.0.4
terminado==0.5
texttable==0.8.3
tornado==4.1
virtualenv==13.1.0
virtualenvwrapper-win==1.2.1
websocket-client==0.32.0
Werkzeug==0.10.4
whitenoise==1.0.6
